pyproject.toml
[project]
name = "test_poetry_cicd"
version = "0.1.0"
description = ""
authors = [
    {name = "Arya-Chakraborty",email = "aryastlawrence@gmail.com"},
    {name = "auhonab", email = "auhonabasu03@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas (>=2.2.3,<3.0.0)",
    "transformers (>=4.52.4,<5.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "fastapi-jwt-auth (>=0.5.0,<0.6.0)",
    "pydantic (>=2.11.5,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "cassandra-driver (>=3.28.0,<4.0.0)",
]

[tool.poetry]
packages = [{include = "test_poetry_cicd", from = "src"}]


[tool.poetry.group.dev.dependencies]
pylint = "^3.3.7"
flake8 = "^7.2.0"

[tool.flake8]
max-line-length = 88
exclude = ["build", "dist", "__pycache__", ".venv", "venv", "migrations"]
ignore = ["E203", "W503"]

[tool.pylint."MESSAGES CONTROL"]
disable = [
    "missing-docstring",
    "invalid-name",
    "too-few-public-methods",
    "too-many-arguments",
    "too-many-instance-attributes"
]

[tool.pylint.FORMAT]
max-line-length = 88
indent-string = "    "

[tool.pylint.DESIGN]
max-args = 10
max-locals = 15
max-branches = 12
max-statements = 50
max-parents = 7
max-attributes = 10
max-public-methods = 20

[tool.pylint.REPORTS]
output-format = "text"
reports = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
